requirements.txt
certifi
chardet
defusedxml
Django
django-allauth
django-haystack
django-jet
django-js-asset
django-machina
django-mptt
django-tinymce4-lite
django-widget-tweaks
idna
jsmin
markdown2
oauthlib
Pillow
pyenchant
python3-openid
pytz
requests
requests-oauthlib
sqlparse
urllib3
django-recaptcha
django-autoslug
django-allauth-2fa